squash! test: improve performance of test-crypto-timing-safe-equal-benchmarks
Use eval() instead....
