+++ test/pummel/test-crypto-timing-safe-equal-benchmarks.js
@@ -6,25 +6,24 @@ if (!common.hasCrypto)
 if (!common.enoughTestMem)
   common.skip('memory-intensive test');
 
-const fixtures = require('../common/fixtures');
 const assert = require('assert');
 const crypto = require('crypto');
 
-const BENCHMARK_FUNC_PATH =
-  `${fixtures.fixturesDir}/crypto-timing-safe-equal-benchmark-func`;
-function runOneBenchmark(...args) {
-  const benchmarkFunc = require(BENCHMARK_FUNC_PATH);
-  const result = benchmarkFunc(...args);
+function runOneBenchmark(compareFunc, firstBufFill, secondBufFill, bufSize) {
+  return eval(`
+      const firstBuffer = Buffer.alloc(bufSize, firstBufFill);
+      const secondBuffer = Buffer.alloc(bufSize, secondBufFill);
 
-  // Don't let the comparison function get cached. This avoid a timing
-  // inconsistency due to V8 optimization where the function would take
-  // less time when called with a specific set of parameters. This used to use
-  // `delete` instead of `Object.create(null)` but that resulted in many times
-  // worse performance, so compare the runtime for this test if you change it
-  // back or otherwise modify the `Object.create(null)` line below.
-  // Ref: https://github.com/nodejs/node/pull/26237
-  require.cache = Object.create(null);
-  return result;
+      const startTime = process.hrtime();
+      const result = compareFunc(firstBuffer, secondBuffer);
+      const endTime = process.hrtime(startTime);
+
+      // Ensure that the result of the function call gets used, so it doesn't
+      // get discarded due to engine optimizations.
+      assert.strictEqual(result, firstBufFill === secondBufFill);
+
+      endTime[0] * 1e9 + endTime[1];
+    `);
 }
 
 function getTValue(compareFunc) {
